update struct KDPendaftar

+++ model/type.go
@@ -6,8 +6,8 @@ import (
 
 type Pendaftaran struct {
 	ID               primitive.ObjectID `bson:"_id,omitempty" json:"_id,omitempty"`
-	KDPendaftar  	 int                `bson:"kdpendaftar,omitempty" json:"kdpendaftar,omitempty"`
-	StatusPendaftar  string                `bson:"statuspendaftar,omitempty" json:"statuspendaftar,omitempty"`
+	KDPendaftar  	 int64              `bson:"kdpendaftar,omitempty" json:"kdpendaftar,omitempty"`
+	StatusPendaftar  string             `bson:"statuspendaftar,omitempty" json:"statuspendaftar,omitempty"`
 	Biodata      	 Camaba             `bson:"biodata,omitempty" json:"biodata,omitempty"`
 	AsalSekolah  	 DaftarSekolah      `bson:"asalsekolah,omitempty" json:"asalsekolah,omitempty"`
 	Jurusan  	 	 Jurusan  	        `bson:"jurusan,omitempty" json:"jurusan,omitempty"`
